reorganize map view controller
